feat(backend): implement OrderMilestoneService for milestone updates and delivery verification
